requirements.txt
python-digitalocean
boto
ansible
